// enums/restaurants.enum.ts
export enum RESTAURANT_TYPE {
  VEG = "veg",
  NON_VEG = "non-veg",
}

export enum RESTAURANT_MODE {
  ONLINE = "online",
  DINEIN = "dinein",
}

export enum ORDER_STATUS {
  PREPARING = 1,
  READY_FOR_PICKUP = 2,
  OUT_FOR_DELIVERY = 3,
  DELIVERED = 4,
  CANCELLED = 5,
  ORDER_PLACED = 6,
  ORDER_FAILED = 7,
}